Fixed rendering issue on GitHub
The initial Jupyter Notebook file weighed nearly 17MB due to its comprehensive inclusion of statistical analyses and visualizations for every year of the Tampa Bay Lightning. This substantial size hindered its ability to be rendered for preview on GitHub. To address this and ensure the notebook serves as a suitable portfolio showcase, I introduced a straightforward conditional check. This check currently limits the rendering to data from the first year. However, it's important to note that this condition should be removed to facilitate the visualization of data for each subsequent year.